adding 404

+++ blog/views.py
@@ -3,6 +3,7 @@ from django.utils import timezone
 from .models import Post
 from .forms import PostForm
 from django.shortcuts import render, get_object_or_404, redirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 
 # Create your views here.
 def post_list(request):
@@ -16,10 +17,14 @@ def post_list(request):
     return render(request, 'blog/post_list.html', {'posts': posts})
 
 def post_detail(request, pk):
+    if not request.user.is_staff:
+        raise Http404
     post = get_object_or_404(Post, pk=pk)
     return render(request, 'blog/post_detail.html', {'post': post})
 
 def post_new(request):
+    if not request.user.is_staff:
+        raise Http404
     if request.method == "POST":
         form = PostForm(request.POST)
         if form.is_valid():
@@ -34,6 +39,8 @@ def post_new(request):
     return render(request, 'blog/post_edit.html', {'form': form})
 
 def post_edit(request, pk):
+    if not request.user.is_staff:
+        raise Http404    
     post = get_object_or_404(Post, pk=pk)
     if request.method == "POST":
         form = PostForm(request.POST, instance=post)
@@ -49,16 +56,22 @@ def post_edit(request, pk):
     return render(request, 'blog/post_edit.html', {'form': form})
 
 def post_draft_list(request):
+    if not request.user.is_staff:
+        raise Http404    
     posts = Post.objects.filter(published_date__isnull=True)\
     .order_by('created_date')
     return render(request, 'blog/post_draft_list.html', {'posts': posts})
 
 def post_publish(request, pk):
+    if not request.user.is_staff:
+        raise Http404    
     post = get_object_or_404(Post, pk=pk)
     post.publish()
     return redirect('post_detail', pk=pk)
 
 def post_remove(request, pk):
+    if not request.user.is_staff:
+        raise Http404    
     post = get_object_or_404(Post, pk=pk)
     post.delete()
     return redirect('post_list')
